backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# Pydantic and Settings
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML Libraries
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
transformers==4.36.0
openai==1.3.7
whisper==1.1.10

# Computer Vision
opencv-python==4.8.1.78
dlib==19.24.2
mediapipe==0.10.8
face-recognition==1.3.0

# Audio Processing
librosa==0.10.1
soundfile==1.0.1
pyaudio==0.2.11

# WebRTC and Real-time Communication
aiortc==1.6.0
websockets==12.0

# Task Queue
celery==5.3.4
flower==2.0.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0

# Data Processing
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Environment and Configuration
python-dotenv==1.0.0
click==8.1.7

# Date and Time
python-dateutil==2.8.2

# JSON and Data Validation
jsonschema==4.20.0

# Code Execution Security
docker==6.1.3
subprocess32==3.5.4

# Natural Language Processing
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1

# Utilities
requests==2.31.0
urllib3==2.1.0
typing-extensions==4.8.0
